refactor: 네비게이션 로직 훅으로 분리
